Refactored & Cleaned up code.
