+++ services/passport.js
@@ -1,0 +1,52 @@
+// -----------------------------------------------------------------------------------------
+// External Dependencies
+// -----------------------------------------------------------------------------------------
+const LocalStrategy = require('passport-local');
+const ExtractJwt    = require('passport-jwt').ExtractJwt;
+const JwtStrategy   = require('passport-jwt').Strategy;
+const passport      = require('passport');
+
+// -----------------------------------------------------------------------------------------
+// Internal Dependencies
+// -----------------------------------------------------------------------------------------
+const User   = require('../models/user');
+const keys   = require('../config/keys');
+
+// -----------------------------------------------------------------------------------------
+// JWT Strategy
+// -----------------------------------------------------------------------------------------
+const jwtOptions = {
+  jwtFromRequest: ExtractJwt.fromHeader('authorization'),
+  secretOrKey: keys.jwtSecret   
+};
+
+const jwtLogin = new JwtStrategy(jwtOptions, (payload, done) => {
+  User.findById(payload.sub, (err, user) => {
+    if (err) return done(err, false);
+
+    if (user) done(null, user);
+    else      done(null, false);
+  });
+});
+
+passport.use(jwtLogin);
+
+// -----------------------------------------------------------------------------------------
+// Local Strategy
+// -----------------------------------------------------------------------------------------
+const localOptions = { usernameField: 'email' };
+
+const localLogin = new LocalStrategy(localOptions, (email, password, done) => {
+  User.findOne({ email: email }, (err, user) => {
+    if (err)   return done(err);
+    if (!user) return done(null, false);
+
+    user.comparePassword(password, (err, isMatch) => {
+      if (err)      return done(err);
+      if (!isMatch) return done(null, false);
+      return done(null, user);
+    });
+  });
+});
+
+passport.use(localLogin);
